% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_project_obs.R
\name{download_many_projects}
\alias{download_many_projects}
\title{Download observations from a vector of project slugs}
\usage{
download_many_projects(x)
}
\arguments{
\item{x}{vector of project slugs}
}
\value{
sf dataframe of inat observations
}
\description{
Download observations from a vector of project slugs
}
\examples{
\dontrun{
projects_vec <- c(
  "city-nature-challenge-2022-greater-vancouver",
  "city-nature-challenge-2022-delta-bc"
)
obs_sf <- download_many_projects(projects_vec)
}
}
