% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_project_obs.R
\name{download_project}
\alias{download_project}
\title{Download all inat observations from a single project}
\usage{
download_project(x)
}
\arguments{
\item{x}{slug identifying a particular project}
}
\value{
sf dataframe of inat observations
}
\description{
Download all inat observations from a single project
}
\examples{
\dontrun{
obs_sf <- download_project("city-nature-challenge-2022-greater-vancouver")
}
}
